+++ library/inspector/src/main/java/io/mehow/laboratory/inspector/LaboratoryActivity.kt
@@ -4,20 +4,21 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.View.OVER_SCROLL_NEVER
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.widget.AppCompatImageView
 import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.core.view.isVisible
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.widget.ViewPager2
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.snackbar.Snackbar
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
 import com.willowtreeapps.hyperion.plugin.v1.HyperionIgnore
 import io.mehow.laboratory.FeatureFactory
 import io.mehow.laboratory.Laboratory
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 
 /**
@@ -25,6 +26,8 @@ import kotlinx.coroutines.launch
  */
 @HyperionIgnore // https://github.com/willowtreeapps/Hyperion-Android/issues/194
 public class LaboratoryActivity : AppCompatActivity(R.layout.io_mehow_laboratory_inspector) {
+  private val searchViewModel by viewModels<SearchViewModel> { SearchViewModel.Factory }
+
   override fun onCreate(inState: Bundle?) {
     super.onCreate(inState)
     setUpToolbar()
@@ -32,10 +35,14 @@ public class LaboratoryActivity : AppCompatActivity(R.layout.io_mehow_laboratory
   }
 
   private fun setUpToolbar() {
-    val resetDialog = createResetDialog()
-    findViewById<AppCompatImageView>(R.id.io_mehow_laboratory_reset_features).setOnClickListener {
-      resetDialog.show()
-    }
+    val binding = ToolbarBinding(
+        view = window.decorView,
+        onSearchEventsListener = { event -> searchViewModel.sendEvent(event) },
+        onResetEventsListener = { resetFeatureFlags() },
+    )
+    searchViewModel.uiModels
+        .onEach { uiModel -> binding.render(uiModel) }
+        .launchIn(lifecycleScope)
   }
 
   private fun setUpViewPager() {
@@ -52,13 +59,6 @@ public class LaboratoryActivity : AppCompatActivity(R.layout.io_mehow_laboratory
       tab.text = sectionNames[position]
     }.attach()
   }
-
-  private fun createResetDialog() = MaterialAlertDialogBuilder(this)
-      .setTitle(R.string.io_mehow_laboratory_reset_title)
-      .setMessage(R.string.io_mehow_laboratory_reset_message)
-      .setNegativeButton(R.string.io_mehow_laboratory_cancel) { _, _ -> }
-      .setPositiveButton(R.string.io_mehow_laboratory_reset) { _, _ -> resetFeatureFlags() }
-      .create()
 
   private fun resetFeatureFlags() = lifecycleScope.launch {
     val isCleared = configuration.laboratory.clear()